Shows a count down view for refresh date
