+++ Example/SnabblePayExample/Views/AddFirstAccountView.swift
@@ -1,0 +1,30 @@
+//
+//  AddFirstAccountView.swift
+//  SnabblePayExample
+//
+//  Created by Uwe Tilemann on 04.04.23.
+//
+
+import SwiftUI
+import SnabblePay
+
+struct AddFirstAccount: View {
+    @ObservedObject var viewModel: AccountsViewModel
+
+    var body: some View {
+        VStack(spacing: 10) {
+            Text("Add your first account now!")
+                .frame(maxWidth: .infinity)
+                .font(.title3)
+                .foregroundColor(.white)
+                .shadow(radius: 2)
+            Button(action: {
+                viewModel.startAccountCheck()
+            }) {
+                Image(systemName: "plus")
+                    .font(.system(size: 64))
+            }
+        }
+        .cardStyle()
+    }
+}
